+++ R/ObjectiveFSelect.R
@@ -82,7 +82,9 @@ ObjectiveFSelect = R6Class("ObjectiveFSelect",
       tasks = map(private$.xss, function(x) {
         state = self$task$feature_names[unlist(x)]
         task = self$task$clone()
-        task$select(state)
+        always_included = task$col_roles$always_included
+        task$set_col_roles(always_included, "feature")
+        task$select(c(state, always_included))
         task
       })