+++ R/zzz.R
@@ -16,6 +16,10 @@
   x = utils::getFromNamespace("bbotk_reflections", ns = "bbotk")
   x$optimizer_properties = c(x$optimizer_properties, "requires_model")
 
+  x = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
+  x$task_col_roles$classif = c(x$task_col_roles$classif, "always_included")
+  x$task_col_roles$regr = c(x$task_col_roles$regr, "always_included")
+
   # callbacks
   x = utils::getFromNamespace("mlr_callbacks", ns = "mlr3misc")
   x$add("mlr3fselect.backup", load_callback_backup)