+++ R/FSelectInstanceMultiCrit.R
@@ -99,6 +99,7 @@ FSelectInstanceMultiCrit = R6Class("FSelectInstanceMultiCrit",
         self$objective$task$feature_names[as.logical(x)]
       })
       xdt[, features := list(features)]
+      xdt[, n_features := map(features, length)]
       assert_data_table(xdt)
       assert_names(names(xdt), must.include = self$search_space$ids())
       assert_data_table(ydt)