+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3fselect (development version)
+
 # mlr3fselect 0.11.0
 
 * BREAKING CHANGE: The `method` parameter of `fselect()`, `fselect_nested()` and `auto_fselector()` is renamed to `fselector`.