+++ NEWS.md
@@ -1,5 +1,8 @@
 # mlr3fselect (development version)
 
+* refactor: Change default batch size of `FSelectorRandomSearch` to 10.
+* feat: Add `batch_size` parameter to `FSelectorExhaustiveSearch` to reduce memory consumption.
+
 # mlr3fselect 0.11.0
 
 * BREAKING CHANGE: The `method` parameter of `fselect()`, `fselect_nested()` and `auto_fselector()` is renamed to `fselector`.