+++ NEWS.md
@@ -1,5 +1,6 @@
 # mlr3fselect (development version)
 
+* feat: Add number of features to `instance$result`.
 * feat: Add `ties_method` options `"least_features"` and `"random"` to `ArchiveFSelect$best()`.
 * refactor: Optimize runtime of `ArchiveFSelect$best()` method.
 * feat: Add importance scores to result of `FSelectorRFE`.