+++ NEWS.md
@@ -10,6 +10,7 @@
 * fix: Include `FSelector` in hash of  `AutoFSelector`.
 * refactor: Change default batch size of `FSelectorRandomSearch` to 10.
 * feat: Add `batch_size` parameter to `FSelectorExhaustiveSearch` to reduce memory consumption.
+* Compatibility with upcoming paradox release.
 
 # mlr3fselect 0.11.0