+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3fselect (development version)
+
 # mlr3fselect 0.12.0
 
 * feat: Add number of features to `instance$result`.